Next iteration
I removed everything unnecessary that was needed for the tests. The algorithm outputs the residual that corresponds to the system with the highest refractive index. In this case, I assume that the refractive indices are given in ascending order. I will assume that the system with the highest refractive index converges last (generally speaking, this is highly debatable).